% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{ari_fixed_agerel}
\alias{ari_fixed_agerel}
\title{Bayesian linear regression with Stan}
\usage{
ari_fixed_agerel(
  N,
  K,
  T,
  ObsT,
  AgeTops,
  R,
  ari_mu = -4,
  ari_sig = 0.7,
  rho_mu = -3,
  rho_sig = 0.2,
  alpha_mu = 0,
  alpha_sig = 0.05,
  ...
)
}
\arguments{
\item{N}{denominator matrix NT_obs x NAge_obs}

\item{K}{numerator matrix NT_obs x NAge_obs}

\item{ObsT}{times of observations (length NAge_obs)}

\item{AgeTops}{locations of age tops}

\item{R}{relative foi by age (length A=number of underlying ages)}

\item{ari_mu}{ARI prior log normal mu}

\item{ari_sig}{ARI prior log normal sigma}

\item{rho_mu}{Reversion prior log normal mu}

\item{rho_sig}{Reversion prior log normal sigma}

\item{alpha_mu}{ARI trend prior normal mu}

\item{alpha_sig}{ARI trend prior normal sigma}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}

\item{b}{per capita birth rate (length T=number of times)}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Bayesian linear regression with Stan
}
\examples{

# run inference on test_data using 4 chains in parallel
out <- ari_fixed_agerel(N=test_data$N,#denominator matrix NT_obs x NAge_obs
K=test_data$K,#numerator matrix NT_obs x NAge_obs
T=test_data$T, #number of times
ObsT=test_data$ObsT, #times of observations
AgeTops=test_data$AgeTops,#locations of age tops
R=test_data$R,            #relative ARI by age
iter=1e3,chains=4,cores=4)

# inspect parameters of interest
prz <- c('lambda0','rho','alpha')
summary(samps,pars=prz)


}
